requirements.txt
kivy==2.2.1
kivy_garden_mapview==1.0.6
pillow==11.3.0
